Extract ConfigParams interface to separate file
