+++ src/ArgumentSanitization.ts
@@ -3,13 +3,14 @@
  * Copyright (c) 2024 Handsoncode. All rights reserved.
  */
 
-import {Config, ConfigParamsList} from './Config'
+import {Config} from './Config'
 import {
   ConfigValueTooBigError,
   ConfigValueTooSmallError,
   ExpectedOneOfValuesError,
   ExpectedValueOfTypeError
 } from './errors'
+import {ConfigParamsList} from './ConfigParams'
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 export function configValueFromParam(inputValue: any, expectedType: string | string[], paramName: ConfigParamsList) {
